+++ parser_library/src/workspace_manager.cpp
@@ -22,7 +22,6 @@
 #include <string_view>
 #include <vector>
 
-#include "nlohmann/json.hpp"
 #include "utils/resource_location.h"
 #include "workspace_manager_impl.h"
 #include "workspace_manager_response.h"
@@ -86,10 +85,9 @@ void workspace_manager::did_close_file(const char* document_uri)
     impl_->did_close_file(utils::resource::resource_location(document_uri));
 }
 
-void workspace_manager::configuration_changed(const lib_config& new_config, const char* whole_settings)
+void workspace_manager::configuration_changed(const lib_config& new_config)
 {
-    impl_->configuration_changed(
-        new_config, std::make_shared<const nlohmann::json>(nlohmann::json::parse(std::string_view(whole_settings))));
+    impl_->configuration_changed(new_config);
 }
 
 void workspace_manager::register_diagnostics_consumer(diagnostics_consumer* consumer)
@@ -113,6 +111,11 @@ void workspace_manager::unregister_parsing_metadata_consumer(parsing_metadata_co
 }
 
 void workspace_manager::set_message_consumer(message_consumer* consumer) { impl_->set_message_consumer(consumer); }
+
+void workspace_manager::set_request_interface(workspace_manager_requests* requests)
+{
+    impl_->set_request_interface(requests);
+}
 
 void workspace_manager::definition(const char* document_uri, position pos, workspace_manager_response<position_uri> r)
 {