+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -13,6 +13,7 @@
 - Inconsistent identification of inactive statements
 - Closing a dependency without saving does not trigger reparsing
 - Source code colorization may be flickering while typing
+- Configuration request sent before initialization is done
 
 ## [1.7.0](https://github.com/eclipse-che4z/che-che4z-lsp-for-hlasm/compare/1.6.0...1.7.0) (2023-03-08)